fix: extract create bundle component
